FIN-28 Receivables refactoring - DateInput reusable component
