+++ src/src_V2/modules/manager/financial/components/index.js
@@ -1,1 +1,2 @@
-export * from './search-input/SearchInput';
+export { default as SearchInput } from './search-input/SearchInput';
+export { default as DateInput } from './date-input/DateInput';
