+++ src/src_V2/modules/manager/financial/payables/edit/drawer-content/editable-info/EditableInfo.module.css
@@ -7,19 +7,3 @@
   width: 25.3vw;
   margin-right: 10px !important;
 }
-
-.dueDatePickerInput {
-  margin-left: 10px;
-}
-
-.dueDatePickerInput > div > div {
-  width: 25.3vw;
-}
-
-.dueDatePickerInput > div > label {
-  top: -6px;
-}
-
-.datePickerLblFocused {
-  top: 0 !important;
-}
