+++ src/src_V2/modules/manager/financial/receivables/edit/drawer-content/editable-info/EditableInfo.jsx
@@ -1,22 +1,14 @@
-import { DatePicker } from '@mui/lab';
-import { Box, TextField } from '@mui/material';
-
+import { DateInput } from '../../../../components';
 import commonStyles from '../style.module.css';
-import styles from './EditableInfo.module.css';
 
 export default function EditableInfo({ values, setFieldValue }) {
   return (
-    <Box className={styles.editableFields}>
-      <Box className={`${commonStyles.dueDatePicker} ${styles.dueDatePickerInput}`}>
-        <DatePicker
-          name='dueDate'
-          label='Due date'
-          inputFormat='dd/MM/yyyy'
-          value={values.dueDate}
-          onChange={(newValue) => setFieldValue('dueDate', newValue?.toString())}
-          renderInput={(params) => <TextField {...params} />}
-        />
-      </Box>
-    </Box>
+    <DateInput
+      name='dueDate'
+      value={values.dueDate}
+      label='Due Date'
+      className={commonStyles.dueDatePickerInput}
+      searchHandler={(value) => setFieldValue('dueDate', value)}
+    />
   );
 }
