--- src/src_V2/modules/manager/financial/receivables/edit/drawer-content/editable-info/EditableInfo.module.css
@@ -1,17 +1,0 @@
-.editableFields {
-  display: flex;
-  justify-content: space-between;
-}
-
-.actualCostControl {
-  width: 25.3vw;
-}
-
-.actualCostControl > div {
-  display: flex;
-  align-items: baseline;
-}
-
-.dueDatePickerInput > div > div {
-  width: 25.3vw;
-}
